Porting Ballistic interface/classes
